pyproject.toml
[tool.poetry]
name = "tradebox"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pydantic = "^2.7.4"
#binance-connector = "^3.7.0"
requests = "^2.32.3"
pydantic-settings = "^2.3.3"
binance-futures-connector = "^4.0.0"
sqlmodel = "^0.0.19"
aiosqlite = "^0.20.0"
greenlet = "^3.0.3"
asyncio = "^3.4.3"
#prefect = "^2.19.6"
tqdm = "^4.66.4"
streamlit = "^1.36.0"
sqladmin = "^0.18.0"
click = "^8.1.7"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
sentry-sdk = { extras = ["fastapi"], version = "^2.10" }
psycopg2 = "^2.9.9"



[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
